pyproject.toml
[project]
name = "kady"
version = "0.1.0"
description = "Kady — K-Dense BYOK Agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.133.1",
    "google-adk[extensions]>=1.27.2",
    "litellm[proxy]>=1.82.3",
    "markitdown[all]>=0.1.5",
    "modal>=1.3.5",
    "openrouter>=0.7.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]
